refactor (core) separate sampling into its own handler
